Make max-results configurable
The one bit I'm uncertain of here is in `query-processor.middleware.limit/determine-query-max-rows`.

So normally when we run a query/question, we'll run middleware that will attach `:max-results` and
`:max-results-bare-rows` to the query. These will correspond to the `max-aggregated-query-row-limit` and
`max-unaggregated-query-row-limit` settings, respectively.

So for questions, we want to use this value.

When we are downloading the raw results of a card, the API layer explicitly removes the `constraints` middleware,
which means `:max-results` and `:max-results-bare-rows` won't be set, which means `(query->max-rows-limit query)`
returns `nil` (assuming the query doesn't have any built-in limits).

In that case, we want to fall back to the `max-unaggregated` or `max-aggregated` query row limits, as appropriate. I
added a new function in the `constraints` middleware ns to allow us.