Skip excess keys on serialization load
When loading from a backup, we should not fail if the previous version saved keys that are no longer needed in the
current version. This adds a step (after ingestion but before loading) that removes any "extra" keys from the ingested
entity.

Co-authored by: bryan <343288+escherize@users.noreply.github.com>