WIP: fix tests for loading with extra keys
This is not the ideal way to fix things, but I'm committing for now as a milepost.

There were two problems with the existing code:

- running the test left the database in a state that caused subsequent tests to fail :sob:

- running the test _without_ first running `dev/start!` would fail

This led to ... extreme confusion, as I'd try to rerun other failing tests and they'd pass, then try running these
tests and they'd fail... or sometimes pass, depending on if I'd run `start!` from my REPL or not... and then try
running the original failing tests again and they'd fail. Just a lot of degrees of freedom, leading to confusion.

The code here is not ideal, obviously - we're digging into the internals of the startup process to put the database
into the state we expect. I'll look into this more tomorrow.